gitlab-runner: Fix empty gitlab_runner_version evaluation in version test
Fix condition evaluation order to check if gitlab_runner_version is empty
before using it in the version test. This prevents the "Input version value
cannot be empty" error when the variable is not set.

The version test was being evaluated before the length check, causing Ansible
to fail when gitlab_runner_version was empty. By checking length first, we
use short-circuit evaluation to skip the version test when the variable is
empty.

Signed-off-by: Norman Ziegner <n.ziegner@hzdr.de>